Extract behaviour from TaskJob to a Concern
