+++ lib/maintenance_tasks/engine.rb
@@ -12,7 +12,7 @@ module MaintenanceTasks
     end
 
     config.to_prepare do
-      _ = TaskJob # load TaskJob for JobIteration compatibility check
+      _ = TaskJobConcern # load this for JobIteration compatibility check
       unless Rails.autoloaders.zeitwerk_enabled?
         tasks_module = MaintenanceTasks.tasks_module.underscore
         Dir["#{Rails.root}/app/tasks/#{tasks_module}/*.rb"].each do |file|
